feat: rename module
