+++ packages/dart_flutter_common/lib/src/widgets/_export.dart
@@ -1,7 +1,6 @@
 export 'generic_chat_room.dart';
 export 'generic_image.dart';
-export 'material_horizontal_card.dart';
-export 'material_vertical_card.dart';
+export 'material_card.dart';
 export 'section.dart';
 export 'selectable_chips.dart';
 export 'web_link.dart';
